--- src/global/ApolloLinkState/namespace.ts
@@ -1,34 +1,0 @@
-import { WithTypeName } from 'typings';
-import * as Yup from 'yup';
-
-import { ImageRef } from 'components/CropperWidget/CropperWidget';
-
-// make optional 'avatarUpload' 'backgroundUpload', change current to strings
-export const cardSchema = Yup.object({
-  name: Yup.string().required(),
-  description: Yup.string().required(),
-  avatar: Yup.mixed<ImageRef>().nullable(true),
-  background: Yup.mixed<ImageRef>().nullable(true),
-});
-
-export const pageSchema = Yup.object({
-  content: Yup.mixed()
-    .required()
-    .nullable(true),
-  avatar: Yup.mixed<ImageRef>().nullable(true),
-  background: Yup.mixed<ImageRef>().nullable(true),
-});
-
-export type CardType = Yup.InferType<typeof cardSchema>;
-
-export type PersonaCard = WithTypeName & CardType;
-
-export type PageType = Yup.InferType<typeof pageSchema>;
-
-export type PersonaPage = WithTypeName & PageType;
-
-export type Persona = WithTypeName & {
-  uuid: string;
-  card: PersonaCard;
-  page: PersonaPage;
-};
