+++ src/global/graphqls/User.ts
@@ -1,10 +1,8 @@
 import { gql } from 'apollo-boost';
 
-//no more condition
-
 export const GET_USER = gql`
-  query user($condition: UserInput!) {
-    user(condition: $condition) {
+  {
+    user {
       uuid
       email
       photo
