+++ src/pages/CreatePersona/CreateCard/CreateCard.tsx
@@ -9,7 +9,7 @@ import { TopNav } from 'components/TopNav/TopNav';
 import { WideButton } from 'components/Button/WideButton';
 import { PageWrapperSpaceBetween } from 'components/PageWrapper';
 import { InfoCard } from 'components/InfoCard/InfoCard';
-import { CardType, cardSchema } from 'global/ApolloLinkState/namespace';
+import { CardType, cardSchema } from 'global/graphqls/schema';
 import { Card } from 'components/Card/Card';
 import { BackgroundPlaceholder } from 'components/BackgroundPlaceholder/BackgroundPlaceholder';
 import { PersonaCircle, PersonaCircleWrapper } from 'components/PersonaCircle/PersonaCircle';
@@ -25,8 +25,8 @@ import { onAvatarChangeHelper, onBgChangeHelper } from '../helpers';
 const cardInitialValues: CardType = {
   name: '',
   description: '',
-  avatar: null,
-  background: null,
+  avatar: '',
+  background: '',
 };
 
 const initialState: ImageRef = {
@@ -69,8 +69,7 @@ export const CreateCard: FC = () => {
   };
 
   const onCrop = (data: ImageRef): void => {
-    setFieldValue(data.fieldName, data, true);
-    // setFieldValue(mapper[data.fieldName], data.blobUrl, true);
+    setFieldValue(data.fieldName, data.blobUrl, true);
     setImageRef(initialState);
   };
 
@@ -93,10 +92,10 @@ export const CreateCard: FC = () => {
             short description.
           </InfoCard>
           <Card mt={31} mb={40} position="relative">
-            <BackgroundPlaceholder background={background?.blobUrl || ''} alt="Card background">
+            <BackgroundPlaceholder background={background || ''} alt="Card background">
               <FileInput onFileChange={onBgChange} name="background" id="background" accept="image/*" />
               <PersonaCircleWrapper>
-                <PersonaCircle avatar={avatar?.blobUrl || ''} alt="Avatar card" onAvatarSet={onAvatarChange} />
+                <PersonaCircle avatar={avatar || ''} alt="Avatar card" onAvatarSet={onAvatarChange} />
               </PersonaCircleWrapper>
             </BackgroundPlaceholder>
             <Flex justifyContent="flex-end" mx={14} my={10}>
